backend/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.0.6
django-rest==0.8.7
django-rest-framework==0.1.0
django-solo==2.0.0
djangorestframework==3.13.1
drf-yasg==1.21.3
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
lxml==4.9.1
MarkupSafe==2.1.1
numpy==1.23.1
packaging==21.3
pandas==1.4.3
prompt-toolkit==3.0.30
psycopg2==2.9.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.10
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
xl